+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+* Added *ordering* wrappers for `Hex` (#147):
+  * `OrdByLength`
+  * `OrdByXY`
+  * `OrdByYX`
+
 ## 0.16.1
 
 * Derived `PartialEq` and `Eq` for `HexBounds` (#160)